Handle RzLang plugins as all the others
