+++ librz/lang/p/pipe_helper.h
@@ -1,0 +1,4 @@
+#include <rz_lang.h>
+#include <rz_types.h>
+
+RZ_IPI int lang_pipe_run(RzLang *lang, const char *code, int len);
